requirements/dev.txt
# This file pulls in everything a developer needs. If it's a basic package
# needed to run the site, it belongs in requirements/prod.txt. If it's a
# package for developers (testing, docs, etc.), it goes in this file.
-r all.txt

# Documentation
Sphinx
#Sphinx==1.1.3

# Testing
coverage
pep8
pylint
django-debug-toolbar

Werkzeug
ipython

django-fiber
-e git+git://github.com/cypreess/django-fiber-modeltranslation.git#egg=fiber_modeltranslation